+++ lib/game/scenes/FirstSequenceScene/FirstSequenceScene.h
